requirements.txt
alembic
aiosqlite
asyncpg
fastapi
httpx
psycopg2-binary
prometheus-client
pydantic
pytest
pytest-asyncio
python-dotenv
sqlalchemy
telethon
uvicorn
selenium
webdriver-manager
beautifulsoup4
lxml
requests

# Arbitrage & Premium Signals Dependencies
firecrawl-py
ccxt
pandas
numpy
websockets
asyncio-throttle
python-binance
coinbase
krakenex
aiohttp
asyncio-mqtt
redis
